Log cleanup, code cleanup

+++ installer/src/registryUtils.cpp
@@ -16,13 +16,13 @@ bool createTaskTrackerKeys() {
 
 	//Base key
 	if (!createRegistryKey(HKEY_CLASSES_ROOT, REGISTRY_PATH, hKey, disposition))
-		return error(L"Failed to create Main Registry key");
+		return logError(L"Failed to create Main Registry key");
 
 	//Base Key Values
 	if (!setRegistryValue(hKey, L"Icon", L"shell32.dll,-4")
 	 || !setRegistryValue(hKey, L"MUIVerb", L"Task Tracker")
 	 || !setRegistryValue(hKey, L"ExtendedSubCommandsKey", REGISTRY_PATH)
-	) return error(L"Failed to set Registry key values");
+	) return logError(L"Failed to set Registry key values");
 
 	const std::array<SubKey, 4> subCommands = {
 		SubKey{L"Finished", FINISHED_PATH, L"C:\\Windows\\System32\\shell32.dll,-16810"},
@@ -36,54 +36,55 @@ bool createTaskTrackerKeys() {
 
 		//Sub Key
 		if (!createRegistryKey(HKEY_CLASSES_ROOT, subCommand.path, subKey, disposition))
-			return log(L"Failed to create Registry subkey: " + subCommand.path.wstring(), true);
+			return logError(L"Failed to create Registry subkey: " + subCommand.path.wstring());
 
 		//Sub Key Values
 		if (!setRegistryValue(subKey, L"MUIVerb", subCommand.name)
 		 || !setRegistryValue(subKey, L"Icon", subCommand.iconPath)
-		) return log(L"Failed to set Registry subkey " + subCommand.path.wstring() + L" values", true);
+		) return logError(L"Failed to set Registry subkey " + subCommand.path.wstring() + L" values");
 
 		//Sub Folder
 		const Path folderPath{ subCommand.path / "command" };
 		if (!createRegistryKey(HKEY_CLASSES_ROOT, folderPath, folderKey, disposition))
-			return log(L"Failed to create Registry subkey " + folderPath.wstring(), true);
+			return logError(L"Failed to create Registry subkey " + folderPath.wstring());
 
 		//Sub Key Command
 		const std::wstring cmd{ L"\"" + getExePath().wstring() + L"\" \"%V\" \"" + subCommand.iconPath + L"\"" };
 		if (!setRegistryValue(folderKey, L"", cmd)) 
-			return log(L"Failed to set Registry subkey " + folderPath.wstring() + L" command", true);
+			return logError(L"Failed to set Registry subkey " + folderPath.wstring() + L" command");
 	}
 
-	return log(L"Program keys successfully added");
+	return logInfo(L"Program keys successfully added\n");
 }
 bool deleteTaskTrackerKeys() {
+	if(!registryKeyExists(REGISTRY_PATH))
+		return logInfo(L"No program keys to delete");
+	
 	bool success{ true };
-
 	success &= deleteRegistryKey(SHELL_PATH);
 	success &= deleteRegistryKey(REGISTRY_PATH);
-
 	return success
-		? log(L"Program Keys successfully removed")
-		: log(L"Partial / Entire failure of program key removal", true);
+		? logInfo(L"Program Keys successfully removed\n")
+		: logError(L"Partial / Entire failure of program key removal\n");
 }
 
 bool registryKeyExists(const std::wstring& path) {
 	return RegOpenKeyExW(HKEY_CLASSES_ROOT, path.data(), 0, KEY_READ, &RegKey{}) != ERROR_SUCCESS
-		? log(L"Key not found at " + path, true) 
-		: log(L"Key found at " + path);
+		? logInfo(L"Key not found at " + path, false) 
+		: logInfo(L"Key found at " + path);
 }
 bool createRegistryKey(const HKEY hKey, const std::wstring& path, RegKey& hKeyHandle, DWORD& disposition) {
 	return RegCreateKeyExW(hKey, path.c_str(), 0, nullptr, 0, KEY_WRITE, nullptr, &hKeyHandle, &disposition) != ERROR_SUCCESS
-		? log(L"Failed to create registry key at " + path, true)
-		: log(L"Created key at " + path);
+		? logError(L"Failed to create registry key at " + path)
+		: logInfo(L"Created key at " + path);
 }
 bool deleteRegistryKey(const std::wstring& path) {
 	return RegDeleteTreeW(HKEY_CLASSES_ROOT, path.c_str()) != ERROR_SUCCESS
-		? log(L"Failed to remove key at " + path, true)
-		: log(L"Removed key at " + path);
+		? logError(L"Failed to remove key at " + path)
+		: logInfo(L"Removed key at " + path);
 }
 bool setRegistryValue(RegKey& hKey, const std::wstring_view& valueName, const std::wstring& value) {
 	return RegSetValueExW(hKey, valueName.data(), 0, REG_SZ, reinterpret_cast<const BYTE*>(value.data()), static_cast<DWORD>((value.size() + 1) * sizeof(wchar_t))) != ERROR_SUCCESS
-		? log(L"Failed to set registry to " + value, true)
-		: log(L"Set registry value to " + value);
+		? logError(L"Failed to set registry to " + value)
+		: logInfo(L"Set registry value to " + value);
 }
